TypingAreaのタイピングロジックをカスタムフックにした
